refactor(metrics): 関数を functions ディレクトリに移動
