+++ src/feature/metric/functions/calculateMaintainabilityIndex.ts
@@ -1,8 +1,8 @@
-import { Tree } from '../../utils/Tree';
-import { CognitiveComplexityMetrics } from './CognitiveComplexity';
-import { CyclomaticComplexityMetrics } from './CyclomaticComplexity';
-import { MetricsScope } from './metricsModels';
-import { SemanticSyntaxVolumeMetrics } from './SemanticSyntaxVolume';
+import { Tree } from '../../../utils/Tree';
+import { CognitiveComplexityMetrics } from '../CognitiveComplexity';
+import { CyclomaticComplexityMetrics } from '../CyclomaticComplexity';
+import { MetricsScope } from '../metricsModels';
+import { SemanticSyntaxVolumeMetrics } from '../SemanticSyntaxVolume';
 
 interface RawMetrics {
   semanticSyntaxVolume: SemanticSyntaxVolumeMetrics;
